backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic[email]>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
python-dotenv>=1.0.1
PyPDF2>=3.0.1
python-docx>=1.1.2
requests>=2.32.0
scikit-learn>=1.5.0
google-generativeai>=0.8.0
google-genai>=1.0.0
groq>=0.4.0
cryptography>=44.0.0
PyJWT>=2.10.0
psycopg2-binary>=2.9.9
sendgrid>=6.10.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.8
pydub>=0.25.1
